Generate cairo-features.h using tup.config, add tup-config-generator.py
Now a tup.config file can be generated using config.status (which needs to be renamed to configstatus.py so python can import it, it seems). The advantage to using tup.config is that tup can then properly detect & update only the parts of the build that are affected by a config change.