Build necko-config.h using tup-preprocessor.py
The preprocessor code is largely cribbed from ConfigStatus.py, but is modified to use the tup_client code to access tup's configuration variables. This way tup can track the dependencies, so for example if the configuration variable NECKO_CONFIG_about is changed, tup will know to re-generate necko-config.h, but it does not need to re-generate all other auto-generated headers.

Initially, an attempt was made to add a '--tup' flag to ConfigStatus.py, so that it could use tup_client if enabled, and default to the current config.status way of doing things otherwise. However, tup's way of providing access to the configuration variables doesn't really have an analog for the 'defines' and 'substs' variables, and things like 'topsrcdir' may change the pathing in unexpected ways. Still, the create_config_header() method could be pulled out in a way that both could share the same code, but that isn't done here.