Install prcpucfg.h.
For some reason this goes in both dist/include and dist/include/nspr. The make build uses a symlink in dist/include/nspr, but makes a copy in dist/include. We just use a symlink for both. This is based on the value of the MDCPUCFG_H variable, which comes from the nspr autoconf.mk file.