Start ipdl conversion.
In the make build, a single python process is executed to read all of the .ipdl files and create the entire ipc/ipdl/_ipdlheaders/ directory structure all at once. With tup, the model is going to be to read a single .ipdl file at a time and write out the contents just for that file. This way a change to a single .ipdl file only rebuilds the pieces necessary, rather than all generated files for every .ipdl file.