Compile xpcom/components.
This one doesn't include config.mk explicitly, and expects it to be included from rules.mk. However, we can't just post-include config.mk since it overwrites CXXFLAGS, which xpcom/components/Makefile.in sets to have the GTK cflags. So we just have to include it whenever rules.mk would normally be included.

Note that we don't always include it, since it doubles the parsing time of dist/include.