expandlibs_exec: Don't try to expand the target itself.
In make builds, expandlibs_exec is called first on libfoo.a, then second on libfoo.a.desc. If we try to build libfoo.a.desc first, then when building libfoo.a, it expands libfoo.a.desc, and tries to write the archive to the first object file in libfoo.a.desc. We should be able to build these two files in any order, or in parallel.